+++ src/Migrations/Creator.php
@@ -12,12 +12,13 @@
 namespace WouterJ\EloquentBundle\Migrations;
 
 use Illuminate\Database\Migrations\MigrationCreator;
+use Symfony\Bundle\MakerBundle\FileManager;
 
 /**
  * A bridge between Illuminate\Database and Symfony.
  *
  * This removes the dependency on Illuminate\Filesystem in favor
- * of PHP's file_*_contents() functions for filesystem tasks.
+ * of MakerBundle's FileManager.
  *
  * @final
  * @internal
@@ -25,21 +26,20 @@ use Illuminate\Database\Migrations\MigrationCreator;
  */
 class Creator extends MigrationCreator
 {
-    // Override constructor to remove Illuminate\Filesystem dep
-    public function __construct()
-    { }
+    private $fileManager;
+
+    public function __construct(FileManager $fileManager)
+    {
+        $this->fileManager = $fileManager;
+    }
 
     /** {@inheritdoc} */
     public function create($name, $path, $table = null, $create = false)
     {
-        if (!is_dir($path)) {
-            mkdir($path, 0777, true);
-        }
-
         $path = $this->getPath($name, $path);
         $stub = $this->getStub($table, $create);
 
-        file_put_contents($path, $this->populateStub($name, $stub, $table));
+        $this->fileManager->dumpFile($path, $this->populateStub($name, $stub, $table));
 
         $this->firePostCreateHooks($table);